feat(core): Implemented ApiError and ApiResponse class along with segregation og author, category

+++ controller/book.controller.js
@@ -1,40 +1,48 @@
 import { Book } from "../model/book.model.js";
 import { z } from "zod";
+import { ApiError } from "../util/ApiError.js";
+import { ApiResponse } from "../util/ApiResponse.js";
 
 const bookSchema = z.object({
-  title: z.string().min(3),
-  description: z.string().min(10),
-  author: z.string().min(3),
-  isbn: z.string().length(13),
-  category: z.string(),
+  title: z.string().min(3, { message: "Title must be at least 3 characters" }),
+  description: z
+    .string()
+    .min(10, { message: "Description must be at least 10 characters" }),
+  authorId: z
+    .string()
+    .min(3, { message: "Author ID must be at least 3 characters" }),
+  isbn: z
+    .string()
+    .length(13, { message: "ISBN must be exactly 13 characters" }),
+  categoryId: z.string(),
 });
 
 const createBook = async (req, res) => {
   try {
-    const result = bookSchema.safeParse(
-      req.body
-    );
-    console.log(result.data)
+    const result = bookSchema.safeParse(req.body);
     if (!result.success) {
-      return res.status(400).json({ errors: result.error.errors });
+      throw new ApiError(400, result.error.errors.join(", "));
     }
-    const { title, description, author, isbn, category } = result.data;
+    const { title, description, authorId, isbn, categoryId, userId } = result.data;
 
-    const book = new Book({ title, description, author, isbn, category });
+    const book = new Book({ title, description, authorId, isbn, categoryId, userId });
     await book.save();
-
-    res.status(201).json({ message: "Book created successfully", book });
+    return res
+      .status(201)
+      .json(new ApiResponse(201, book, "Book created successfully."));
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    throw new ApiError(400, error.message);
   }
 };
 
 const getBooks = async (req, res) => {
   try {
     const books = await Book.find();
-    res.status(200).json(books);
+    return res
+      .status(201)
+      .json(new ApiResponse(200, books, "Books retrieved successfully."));
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    throw new ApiError(400, error.message);
   }
 };
 
@@ -42,45 +50,58 @@ const getBookById = async (req, res) => {
   try {
     const book = await Book.findById(req.params.id);
     if (!book) {
-      return res.status(404).json({ message: "Book not found" });
+      throw new ApiError(404, "Book not found");
     }
-    res.status(200).json(book);
+    return res
+      .status(200)
+      .json(new ApiResponse(200, book, "Book retrieved successfully."));
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    throw new ApiError(400, error.message);
   }
 };
 
 const updateBook = async (req, res) => {
   try {
-    const result = bookSchema.partial().safeParse(req.body);
+    const result = bookSchema.safeParse(req.body);
     if (!result.success) {
-      return res.status(400).json({ errors: result.error.errors });
+      throw new ApiError(400, result.error.errors.join(", "));
     }
 
     const book = await Book.findByIdAndUpdate(req.params.id, result.data, {
       new: true,
       runValidators: true,
     });
+
     if (!book) {
-      return res.status(404).json({ message: "Book not found" });
+      throw new ApiError(404, "Book not found");
     }
 
-    res.status(200).json(book);
+    return res
+      .status(200)
+      .json(new ApiResponse(200, book, "Book updated successfully."));
   } catch (error) {
-    res.status(500).json({ error: error.message });
+    throw new ApiError(400, error.message);
   }
 };
 
 const deleteBook = async (req, res) => {
   try {
-    const book = await Book.findByIdAndDelete(req.params.id);
-    if (!book) {
-      return res.status(404).json({ message: "Book not found" });
+    const { id } = req.params;
+    if (!id) {
+      throw new ApiError(400, "Book ID is required");
     }
 
-    res.status(200).json({ message: "Book deleted successfully" });
+    const book = await Book.findByIdAndDelete(req.params.id);
+
+    if (!book) {
+      throw new ApiError(404, "Book not found");
+    }
+
+    return res
+      .status(200)
+      .json(new ApiResponse(200, "Book deleted successfully."));
   } catch (error) {
-    res.status(500).json({ error: error.message });
+    throw new ApiError(400, error.message);
   }
 };
